services/main/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose
pydantic
python-dotenv
